Glossary/US191T200: I changed my page and sorted the glossary content definitions into their respective six categories using a collapsible list component. I designed it too look exactly like the format from the findHelp pages. I also added a card component to match.
